requirements.txt
requests
PyGithub
python-gitlab